strict checking message reduction

+++ lapack/lapack_utils.c
@@ -1588,7 +1588,7 @@ w;
 	if (iwarn) {
 	    first = TRUE_;
 	    printf("\n\n WARNING. The value EMIN may be incorrect:- ");
-	    printf("EMIN = %8i\n",lemin);
+	    printf("EMIN = %8i\n",(int)lemin);
 	    printf("If, after inspection, the value EMIN looks acceptable");
             printf("please comment out \n the IF block as marked within the"); 
             printf("code of routine DLAMC2, \n otherwise supply EMIN");
